Extract domain-specific validations from Validator and rename it to Parser. The previous parser goes out of the picture.
