+++ README.md
@@ -25,7 +25,6 @@ require 'jsonapi'
 Then, see docs for the part you are interested in:
 * [parsing](parser/README.md)
 * [rendering](renderer/README.md)
-* [validating](validator/README.md)
 
 ## License
