+++ parser/jsonapi-parser.gemspec
@@ -5,13 +5,13 @@ Gem::Specification.new do |spec|
   spec.version       = version
   spec.author        = 'Lucas Hosseini'
   spec.email         = 'lucas.hosseini@gmail.com'
-  spec.summary       = 'Parse and validate JSON API documents'
-  spec.description   = 'Tools for handling JSON API documents'
+  spec.summary       = 'Parse JSONAPI documents'
+  spec.description   = 'Parse JSONAPI response documents, resource ' \
+                       'creation/update payloads, and relationship ' \
+                       'update payloads.'
   spec.homepage      = 'https://github.com/beauby/jsonapi'
   spec.license       = 'MIT'
 
   spec.files         = Dir['README.md', 'lib/**/*']
   spec.require_path  = 'lib'
-
-  spec.add_dependency 'jsonapi-validator'
 end
