+++ parser/lib/jsonapi/parser/resource.rb
@@ -1,25 +1,20 @@
+require 'jsonapi/parser/document'
+
 module JSONAPI
   module Parser
-    # c.f. http://jsonapi.org/format/#document-resource-objects
     class Resource
-      attr_reader :id, :type, :attributes, :relationships, :links, :meta
-
-      def initialize(resource_hash, options = {})
-        @hash = resource_hash
-        @options = options.dup
-        @id = resource_hash['id']
-        @type = resource_hash['type']
-        @attributes_hash = resource_hash['attributes'] || {}
-        @attributes = Attributes.new(@attributes_hash, @options)
-        @relationships_hash = resource_hash['relationships'] || {}
-        @relationships = Relationships.new(@relationships_hash, @options)
-        @links_hash = resource_hash['links'] || {}
-        @links = Links.new(@links_hash, @options)
-        @meta = resource_hash['meta'] if resource_hash.key?('meta')
-      end
-
-      def to_hash
-        @hash
+      # Validate the structure of a resource create/update payload.
+      #
+      # @param [Hash] document The input JSONAPI document.
+      # @raise [JSONAPI::Parser::InvalidDocument] if document is invalid.
+      def self.parse!(document)
+        Document.ensure!(document.is_a?(Hash),
+                         'A JSON object MUST be at the root of every JSONAPI ' \
+                         'request and response containing data.')
+        Document.ensure!(document.key?('data') || !document['data'].is_a?(Hash),
+                         'The request MUST include a single resource object ' \
+                         'as primary data.')
+        Document.parse_primary_resource!(document['data'])
       end
     end
   end
