--- parser/spec/spec_helper.rb
@@ -1,6 +1,0 @@
-require 'simplecov'
-
-# SimpleCov.minimum_coverage 100
-
-SimpleCov.start do
-end
