--- validator/lib/jsonapi/validator/exceptions.rb
@@ -1,6 +1,0 @@
-module JSONAPI
-  module Validator
-    class InvalidDocument < StandardError
-    end
-  end
-end
